+++ apis/academics.ts
@@ -1,13 +1,6 @@
-import {
-  getMockAcademicsGuide,
-  getMockCourseChanges,
-  getMockCourses,
-  getMockGeneralStudiesRequirements,
-  getMockScholarship,
-  getMockUndergraduateScholarshipList,
-} from '@/data/academics';
+import { getMockAcademicsGuide, getMockCourseChanges, getMockCourses } from '@/data/academics';
 
-import { Course, CourseChange, Curriculum, DegreeRequirements, Guide } from '@/types/academics';
+import { Course, CourseChange, Guide } from '@/types/academics';
 
 import { getRequest } from '.';
 
@@ -18,12 +11,6 @@ export const getAcademicsGuide = getMockAcademicsGuide;
 export const getCourses = getMockCourses;
 
 export const getCourseChanges = getMockCourseChanges;
-
-export const getGeneralStudiesRequirements = getMockGeneralStudiesRequirements;
-
-export const getUndergraduateScholarshipList = getMockUndergraduateScholarshipList;
-
-export const getScholarship = getMockScholarship;
 
 // export const getAcademicsGuide = (type: StudentType) =>
 //   getRequest(`/academics/${type}/guide`) as Promise<Guide>;
