+++ app/academics/undergraduate/general-studies-requirements/page.tsx
@@ -1,4 +1,4 @@
-import { getGeneralStudiesRequirements } from '@/apis/academics';
+import { getGeneralStudiesRequirements } from '@/apis/academicsServer';
 
 import HTMLViewer from '@/components/editor/HTMLViewer';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
