+++ data/htmls/scholarship/4.txt
@@ -1,0 +1,1 @@
+<h2 style="margin-top: 1.07143em; margin-bottom: 0.714286em; font-family: &quot;Lucida Sans Unicode&quot;, Verdana, Helvetica, Arial, &quot;맑은 고딕&quot;, 나눔고딕, 서울남산체, sans-serif; color: rgb(109, 172, 255); font-size: 1.4em; padding-bottom: 0.357143em; border-bottom-width: 1px; border-bottom-color: rgb(236, 236, 236); overflow: hidden;">장학 금액</h2><p style="margin-top: 1em; margin-bottom: 1em; color: rgb(51, 51, 51); font-family: Arial, Verdana, Helvetica, sans-serif; font-size: 12px;">등록금 전액 ~ 등록금 10% 면제까지 다양</p><h2 style="margin-top: 1.07143em; margin-bottom: 0.714286em; font-family: &quot;Lucida Sans Unicode&quot;, Verdana, Helvetica, Arial, &quot;맑은 고딕&quot;, 나눔고딕, 서울남산체, sans-serif; color: rgb(109, 172, 255); font-size: 1.4em; padding-bottom: 0.357143em; border-bottom-width: 1px; border-bottom-color: rgb(236, 236, 236); overflow: hidden;">자격</h2><p style="margin-top: 1em; margin-bottom: 1em; color: rgb(51, 51, 51); font-family: Arial, Verdana, Helvetica, sans-serif; font-size: 12px;">학사과정 재학생</p><h2 style="margin-top: 1.07143em; margin-bottom: 0.714286em; font-family: &quot;Lucida Sans Unicode&quot;, Verdana, Helvetica, Arial, &quot;맑은 고딕&quot;, 나눔고딕, 서울남산체, sans-serif; color: rgb(109, 172, 255); font-size: 1.4em; padding-bottom: 0.357143em; border-bottom-width: 1px; border-bottom-color: rgb(236, 236, 236); overflow: hidden;">선발기준</h2><p style="margin-top: 1em; margin-bottom: 1em; color: rgb(51, 51, 51); font-family: Arial, Verdana, Helvetica, sans-serif; font-size: 12px;">학업성취도, 가계곤란, 사회적 배려대상 등을 고려하여 배정</p><h2 style="margin-top: 1.07143em; margin-bottom: 0.714286em; font-family: &quot;Lucida Sans Unicode&quot;, Verdana, Helvetica, Arial, &quot;맑은 고딕&quot;, 나눔고딕, 서울남산체, sans-serif; color: rgb(109, 172, 255); font-size: 1.4em; padding-bottom: 0.357143em; border-bottom-width: 1px; border-bottom-color: rgb(236, 236, 236); overflow: hidden;">신청 시기</h2><ul style="margin-top: 1em; margin-bottom: 1em; padding-left: 2em; list-style-type: square; color: rgb(51, 51, 51); font-family: Arial, Verdana, Helvetica, sans-serif; font-size: 12px;"><li style="margin: 0.2em 0px; padding: 0px;">마이스누 및 학부홈페이지에 공지</li><li style="margin: 0.2em 0px; padding: 0px;">1학기 (직전학년도 11월 하순 ~ 12월 중순)</li><li style="margin: 0.2em 0px; padding: 0px;">2학기 (당해학년도 5월 하순 ~ 6월 초순)</li></ul><h2 style="margin-top: 1.07143em; margin-bottom: 0.714286em; font-family: &quot;Lucida Sans Unicode&quot;, Verdana, Helvetica, Arial, &quot;맑은 고딕&quot;, 나눔고딕, 서울남산체, sans-serif; color: rgb(109, 172, 255); font-size: 1.4em; padding-bottom: 0.357143em; border-bottom-width: 1px; border-bottom-color: rgb(236, 236, 236); overflow: hidden;">신청 절차</h2><p style="margin-top: 1em; margin-bottom: 1em; color: rgb(51, 51, 51); font-family: Arial, Verdana, Helvetica, sans-serif; font-size: 12px;"><a rel="nofollow" href="http://my.snu.ac.kr/" style="color: rgb(27, 125, 236);">서울대 포털 마이스누 접속</a>→ 학사행정 → 등록/장학 → 장학 → [장학복지지원카드]에 작성 → 신청 및 저장 → [장학복지지원카드] 출력 → 학부 행정실에 제출</p>
