+++ data/htmls/scholarship/5.txt
@@ -1,0 +1,1 @@
+<h2 style="margin-top: 1.07143em; margin-bottom: 0.714286em; font-family: &quot;Lucida Sans Unicode&quot;, Verdana, Helvetica, Arial, &quot;맑은 고딕&quot;, 나눔고딕, 서울남산체, sans-serif; color: rgb(109, 172, 255); font-size: 1.4em; padding-bottom: 0.357143em; border-bottom-width: 1px; border-bottom-color: rgb(236, 236, 236); overflow: hidden;">장학 금액</h2><p style="margin-top: 1em; margin-bottom: 1em; color: rgb(51, 51, 51); font-family: Arial, Verdana, Helvetica, sans-serif; font-size: 12px;">등록금 전액 및 수업료 면제</p><h2 style="margin-top: 1.07143em; margin-bottom: 0.714286em; font-family: &quot;Lucida Sans Unicode&quot;, Verdana, Helvetica, Arial, &quot;맑은 고딕&quot;, 나눔고딕, 서울남산체, sans-serif; color: rgb(109, 172, 255); font-size: 1.4em; padding-bottom: 0.357143em; border-bottom-width: 1px; border-bottom-color: rgb(236, 236, 236); overflow: hidden;">자격</h2><p style="margin-top: 1em; margin-bottom: 1em; color: rgb(51, 51, 51); font-family: Arial, Verdana, Helvetica, sans-serif; font-size: 12px;">정원외 외국인특별전형 (외국인), 정원외 (영주자), 외국인편입학 학생 (학사: 2.7 이상, 규정학기 초과자 제외)</p><h2 style="margin-top: 1.07143em; margin-bottom: 0.714286em; font-family: &quot;Lucida Sans Unicode&quot;, Verdana, Helvetica, Arial, &quot;맑은 고딕&quot;, 나눔고딕, 서울남산체, sans-serif; color: rgb(109, 172, 255); font-size: 1.4em; padding-bottom: 0.357143em; border-bottom-width: 1px; border-bottom-color: rgb(236, 236, 236); overflow: hidden;">선발 기준</h2><p style="margin-top: 1em; margin-bottom: 1em; color: rgb(51, 51, 51); font-family: Arial, Verdana, Helvetica, sans-serif; font-size: 12px;">복지과 배정 및 학부 추천</p><h2 style="margin-top: 1.07143em; margin-bottom: 0.714286em; font-family: &quot;Lucida Sans Unicode&quot;, Verdana, Helvetica, Arial, &quot;맑은 고딕&quot;, 나눔고딕, 서울남산체, sans-serif; color: rgb(109, 172, 255); font-size: 1.4em; padding-bottom: 0.357143em; border-bottom-width: 1px; border-bottom-color: rgb(236, 236, 236); overflow: hidden;">신청 시기</h2><ul style="margin-top: 1em; margin-bottom: 1em; padding-left: 2em; list-style-type: square; color: rgb(51, 51, 51); font-family: Arial, Verdana, Helvetica, sans-serif; font-size: 12px;"><li style="margin: 0.2em 0px; padding: 0px;">1학기(직전학년도 11월 중ㆍ하순)</li><li style="margin: 0.2em 0px; padding: 0px;">2학기(당해학년도 5월 하순 ~ 6월 초순)</li></ul><h2 style="margin-top: 1.07143em; margin-bottom: 0.714286em; font-family: &quot;Lucida Sans Unicode&quot;, Verdana, Helvetica, Arial, &quot;맑은 고딕&quot;, 나눔고딕, 서울남산체, sans-serif; color: rgb(109, 172, 255); font-size: 1.4em; padding-bottom: 0.357143em; border-bottom-width: 1px; border-bottom-color: rgb(236, 236, 236); overflow: hidden;">신청 절차</h2><p style="margin-top: 1em; margin-bottom: 1em; color: rgb(51, 51, 51); font-family: Arial, Verdana, Helvetica, sans-serif; font-size: 12px;">별도의 신청절차 없음</p>
