+++ data/htmls/scholarshipList.txt
@@ -1,0 +1,1 @@
+<p><span style="color: rgb(51, 51, 51); font-family: Arial, Verdana, Helvetica, sans-serif; font-size: 12px;">컴퓨터공학부에서는 맞춤형장학금, 근로장학금 등 여러 종류의 교내 장학금과 관정이종환교육재단 등 외부 우수 재단으로부터 다수의 장학금 수혜의 기회를 제공하고 있습니다. 매해 평균 학부생 60% 이상이 교내외 장학금에서 수혜를 받고 있습니다.</span><br></p><h2 style="margin: 1.07143em 0px 0.714286em; font-family: &quot;Lucida Sans Unicode&quot;, Verdana, Helvetica, Arial, &quot;맑은 고딕&quot;, 나눔고딕, 서울남산체, sans-serif; color: rgb(109, 172, 255); font-size: 1.4em; padding-bottom: 0.357143em; border-bottom: 1px solid rgb(236, 236, 236); overflow: hidden; font-style: normal; font-variant-ligatures: normal; font-variant-caps: normal; letter-spacing: normal; orphans: 2; text-align: start; text-indent: 0px; text-transform: none; widows: 2; word-spacing: 0px; -webkit-text-stroke-width: 0px; white-space: normal; background-color: rgb(255, 255, 255); text-decoration-thickness: initial; text-decoration-style: initial; text-decoration-color: initial;">유의사항<em>​</em></h2><ul style="margin: 1em 0px; padding-left: 2em; list-style-type: square; color: rgb(51, 51, 51); font-family: Arial, Verdana, Helvetica, sans-serif; font-size: 12px; font-style: normal; font-variant-ligatures: normal; font-variant-caps: normal; font-weight: 400; letter-spacing: normal; orphans: 2; text-align: start; text-indent: 0px; text-transform: none; widows: 2; word-spacing: 0px; -webkit-text-stroke-width: 0px; white-space: normal; background-color: rgb(255, 255, 255); text-decoration-thickness: initial; text-decoration-style: initial; text-decoration-color: initial;"><li style="margin: 0.2em 0px; padding: 0px;">모든 장학금은 본인신청을 필수로 합니다.</li><li style="margin: 0.2em 0px; padding: 0px;">대부분의 장학금은 소득분위(국가장학금 신청 후 산정)정보가 필요하니 국가장학금을 반드시 신청하시기 바랍니다.<ul style="margin: 0px; padding-left: 2em; list-style-type: circle;"><li style="margin: 0.2em 0px; padding: 0px;">교내장학금 신청시기: 1학기 장학금은 매년 11월~12월, 2학기 장학금은 매년 5월~6월</li><li style="margin: 0.2em 0px; padding: 0px;">교내장학금 신청방법: 마이스누에서 신청 (장학금 신청사유 등 기재)</li><li style="margin: 0.2em 0px; padding: 0px;">교내장학생 선정시기: 1학기 장학생은 2월, 2학기 장학생은 8월</li></ul></li><li style="margin: 0.2em 0px; padding: 0px;">교외장학금: 학부 홈페이지에 공지되며, 재단에 따라 제출서류 및 신청바법 상이</li><li style="margin: 0.2em 0px; padding: 0px;">교장학금 신청방법: 장학금별로 상이하니 홈페이지 공지사항 확인 필수 (모든 장학 서류는 원본 제출)</li><li style="margin: 0.2em 0px; padding: 0px;">장학금은 지급기준에 부합하는 학생을 대상으로 하되, 가정형편을 고려하여 최종 선정하게 됩니다.</li><li style="margin: 0.2em 0px; padding: 0px;">학기별 장학생선정 등 세부사항은<a rel="nofollow" href="https://cse.snu.ac.kr/department-notices?c%5B%5D=1" style="color: rgb(27, 125, 236); text-decoration: none;">공지사항 게시판</a>에 공고됩니다.</li></ul>
