+++ data/serverObjects.ts
@@ -1,7 +1,13 @@
 import { readFileSync } from 'fs';
 import { readFile } from 'fs/promises';
 
-import { Curriculum, DegreeRequirements } from '@/types/academics';
+import {
+  Curriculum,
+  DegreeRequirements,
+  GeneralStudiesRequirements,
+  Scholarship,
+  ScholarshipList,
+} from '@/types/academics';
 import { ResearchLab } from '@/types/research';
 
 import { simpleResearchLabs } from './objects';
@@ -753,4 +759,95 @@ export const getMockDegreeRequirements = async (): Promise<DegreeRequirements[]>
       },
     },
   ];
+};
+
+export const getMockGeneralStudiesRequirements = async (): Promise<GeneralStudiesRequirements> => {
+  return {
+    overview:
+      '컴퓨터공학부 학생이 졸업을 하기 위해 반드시 들어야 하는 영역별 교양과목 학점 배분 구조표입니다. 학부생들은 아래의 구조표를 참고하여 수강에 차질이 없도록 하여야 합니다.',
+    subjectChanges: [
+      {
+        status: 'change',
+        time: 2014,
+        changes: [
+          '(구) 010.001 대학국어(Korean) 3-2-2',
+          '(신) 031.001 글쓰기의 기초(College Writing: Process & Structure) 3-3-0',
+        ],
+        progress: '2014학년도 이전 입학자는 ‘대학국어’ 이수를 ‘글쓰기의 기초’로 대체 이수 가능',
+      },
+      {
+        status: 'change',
+        time: 2010,
+        changes: [
+          '(구) 010.142 컴퓨터의 기초(Fundamental of Computer System) 2-2-0',
+          '(신) 010.133 컴퓨터의 개념 및 실습(Digital Computer Concept and Practice) 3-2-2',
+          '(구) 010.143 컴퓨터원리(Principles of Computer System) 3-3-0',
+          '(신) 010.133 컴퓨터의 개념 및 실습(Digital Computer Concept and Practice) 3-2-2',
+        ],
+        progress:
+          '2009학년도 및 이전 입학자는 ‘컴퓨터의 기초 또는 컴퓨터원리’ 이수를 ‘컴퓨터의 개념 및 실습’으로 대체 이수 가능',
+      },
+      {
+        status: 'new',
+        time: 2005,
+        additionalInfos: [
+          '공과대학 이수규정 변경에 따라 전공교과목에서 교양교과목으로 변경하여 신설하는 것임.',
+          '‘컴퓨터개론(Introduction to Computer)’을 ‘컴퓨터원리’로 명칭변경함.',
+        ],
+        changes: [
+          '010.140 공학수학1(Engineering Mathematics 1) 3-3-0',
+          '010.141 공학수학2(Engineering Mathematics 2) 3-3-0',
+          '010.142 컴퓨터의기초(Fundamentals of Computer System) 2-2-0',
+          '010.143 컴퓨터원리(Principles of Computer System) 3-3-0',
+          '005.054 경영학개론(Introduction to the Theory of Business) 3-3-0',
+        ],
+        progress:
+          '2004학년도 이전 입학자중 전공으로 교양 ‘공학수학1·2’ ‘ 컴퓨터의 기초’ 및 ‘컴퓨터원리’ 교과목을 처음 이수하거나 재이수 하는 경우 전공으로 인정함.',
+      },
+    ],
+    description: await readFile('data/htmls/generalStudiesRequirements.txt', { encoding: 'utf-8' }),
+  };
+};
+
+export const getMockUndergraduateScholarshipList = async (): Promise<ScholarshipList> => {
+  return {
+    description: await readFile('data/htmls/scholarshipList.txt', { encoding: 'utf-8' }),
+    scholarship: [
+      { id: 1, name: '성적우수 국가장학금 (이공계)' },
+      { id: 2, name: '대통령과학장학금' },
+      { id: 3, name: '국가장학금(I,II)' },
+      { id: 4, name: '단대 맞춤형 장학금' },
+      { id: 5, name: '유학생 장학금' },
+      {
+        id: 6,
+        name: '서울대학교 발전기금 장학금(신영길장학금, 현암 김종상장학금, 기타 컴퓨터공학부 지정 장학금 다수)',
+      },
+      {
+        id: 7,
+        name: '교외장학금 (현송문화재단, 유한재단, 관악회,동부문화재단, 청합장학회, 봉신장학회 등 다수)',
+      },
+      { id: 8, name: '컴퓨터공학 동문회 장학금' },
+      { id: 9, name: '근로장학금' },
+    ],
+  };
+};
+
+const scholarshipNames = [
+  '성적우수 국가장학금 (이공계)',
+  '대통령과학장학금',
+  '국가장학금(I,II)',
+  '단대 맞춤형 장학금',
+  '유학생 장학금',
+  '서울대학교 발전기금 장학금(신영길장학금, 현암 김종상장학금, 기타 컴퓨터공학부 지정 장학금 다수)',
+  '교외장학금 (현송문화재단, 유한재단, 관악회,동부문화재단, 청합장학회, 봉신장학회 등 다수)',
+  '컴퓨터공학 동문회 장학금',
+  '근로장학금',
+];
+
+export const getMockScholarship = async (id: number): Promise<Scholarship> => {
+  return {
+    id: id,
+    name: scholarshipNames[id - 1],
+    description: await readFile(`data/htmls/scholarship/${id}.txt`, { encoding: 'utf-8' }),
+  };
 };
